+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## Pending Release
 
 ### Added
-- 
+- Option to check for new version on start up (disabled by default)
 
 ### Changed
 - Moved sorting functions from Lists tool to their own Sort tool.